+++ lib/analysis.dart
@@ -15,3 +15,4 @@ export 'package:jyotish/src/analysis/progeny.dart';
 export 'package:jyotish/src/analysis/progeny_service.dart';
 export 'package:jyotish/src/analysis/yoga.dart';
 export 'package:jyotish/src/analysis/yoga_service.dart';
+export 'package:jyotish/src/analysis/chart_renderer.dart';