+++ lib/jyotish.dart
@@ -57,6 +57,7 @@ export 'package:jyotish/src/analysis/progeny.dart';
 export 'package:jyotish/src/analysis/progeny_service.dart';
 export 'package:jyotish/src/analysis/yoga.dart';
 export 'package:jyotish/src/analysis/yoga_service.dart';
+export 'package:jyotish/src/analysis/chart_renderer.dart';
 export 'package:jyotish/src/analysis/dosha.dart';
 export 'package:jyotish/src/analysis/dosha_service.dart';